refactor : 클래스 HttpBody의 이름을 RequestBody로 수정
- HTTP Response는 Body를 byte[]로 갖습니다.
